+++ src/App.tsx
@@ -33,7 +33,6 @@ export const App: React.FC = () => {
       .then(setTodos)
       .catch(() => {
         setErrorMessage('Unable to load todos');
-        //setTimeout(() => setErrorMessage(''), 3000);
       })
       .finally(() => {
         setIsLoading(false);
@@ -63,7 +62,6 @@ export const App: React.FC = () => {
       setTempTodo(null);
     } catch (error) {
       setErrorMessage('Unable to add a todo');
-      //setTimeout(() => setErrorMessage(''), 3000);
       setTempTodo(null);
     } finally {
       setIsLoading(false);
@@ -72,7 +70,6 @@ export const App: React.FC = () => {
 
   const handleError = (message: string) => {
     setErrorMessage(message);
-    //setTimeout(() => setErrorMessage(''), 3000);
   };
 
   const handleDeleteTodo = async (id: number) => {
@@ -166,8 +163,6 @@ export const App: React.FC = () => {
             );
           }
         });
-
-        //setTimeout(() => setErrorMessage(''), 3000);
       })
       .finally(() => {
         setIsLoading(false);
@@ -197,24 +192,6 @@ export const App: React.FC = () => {
           isAllCompleted={todos.every(todo => todo.completed)}
           onToggleAll={handleToggleAll}
         />
-        {/* <header className="todoapp__header"> */}
-        {/* this button should have `active` class only if all todos are completed */}
-        {/* <button
-            type="button"
-            className="todoapp__toggle-all active"
-            data-cy="ToggleAllButton"
-          /> */}
-
-        {/* Add a todo on form submit */}
-        {/* <form>
-            <input
-              data-cy="NewTodoField"
-              type="text"
-              className="todoapp__new-todo"
-              placeholder="What needs to be done?"
-            />
-          </form>
-        </header> */}
 
         <TodoList
           todos={filteredTodos}
@@ -234,188 +211,12 @@ export const App: React.FC = () => {
             onDeleteCompleted={handleDeleteCompletedTodos}
           />
         )}
-
-        {/* <section className="todoapp__main" data-cy="TodoList"> */}
-        {/* This is a completed todo */}
-        {/* <div data-cy="Todo" className="todo completed">
-            <label className="todo__status-label">
-              <input
-                data-cy="TodoStatus"
-                type="checkbox"
-                className="todo__status"
-                checked
-              />
-            </label>
-
-            <span data-cy="TodoTitle" className="todo__title">
-              Completed Todo
-            </span> */}
-
-        {/* Remove button appears only on hover */}
-        {/* <button type="button" className="todo__remove" data-cy="TodoDelete">
-              ×
-            </button> */}
-
-        {/* overlay will cover the todo while it is being deleted or updated */}
-        {/* <div data-cy="TodoLoader" className="modal overlay">
-              <div className="modal-background has-background-white-ter" />
-              <div className="loader" />
-            </div>
-          </div> */}
-
-        {/* This todo is an active todo */}
-        {/* <div data-cy="Todo" className="todo">
-            <label className="todo__status-label">
-              <input
-                data-cy="TodoStatus"
-                type="checkbox"
-                className="todo__status"
-              />
-            </label>
-
-            <span data-cy="TodoTitle" className="todo__title">
-              Not Completed Todo
-            </span>
-            <button type="button" className="todo__remove" data-cy="TodoDelete">
-              ×
-            </button>
-
-            <div data-cy="TodoLoader" className="modal overlay">
-              <div className="modal-background has-background-white-ter" />
-              <div className="loader" />
-            </div>
-          </div> */}
-
-        {/* This todo is being edited */}
-        {/* <div data-cy="Todo" className="todo">
-            <label className="todo__status-label">
-              <input
-                data-cy="TodoStatus"
-                type="checkbox"
-                className="todo__status"
-              />
-            </label> */}
-
-        {/* This form is shown instead of the title and remove button */}
-        {/* <form>
-              <input
-                data-cy="TodoTitleField"
-                type="text"
-                className="todo__title-field"
-                placeholder="Empty todo will be deleted"
-                value="Todo is being edited now"
-              />
-            </form>
-
-            <div data-cy="TodoLoader" className="modal overlay">
-              <div className="modal-background has-background-white-ter" />
-              <div className="loader" />
-            </div>
-          </div> */}
-
-        {/* This todo is in loadind state */}
-        {/* <div data-cy="Todo" className="todo">
-            <label className="todo__status-label">
-              <input
-                data-cy="TodoStatus"
-                type="checkbox"
-                className="todo__status"
-              />
-            </label>
-
-            <span data-cy="TodoTitle" className="todo__title">
-              Todo is being saved now
-            </span>
-
-            <button type="button" className="todo__remove" data-cy="TodoDelete">
-              ×
-            </button> */}
-
-        {/* 'is-active' class puts this modal on top of the todo */}
-        {/* <div data-cy="TodoLoader" className="modal overlay is-active">
-              <div className="modal-background has-background-white-ter" />
-              <div className="loader" />
-            </div>
-          </div>
-        </section> */}
-
-        {/* Hide the footer if there are no todos */}
-        {/* <footer className="todoapp__footer" data-cy="Footer">
-          <span className="todo-count" data-cy="TodosCounter">
-            3 items left
-          </span> */}
-
-        {/* Active link should have the 'selected' class */}
-        {/* <nav className="filter" data-cy="Filter">
-            <a
-              href="#/"
-              className="filter__link selected"
-              data-cy="FilterLinkAll"
-            >
-              All
-            </a>
-
-            <a
-              href="#/active"
-              className="filter__link"
-              data-cy="FilterLinkActive"
-            >
-              Active
-            </a>
-
-            <a
-              href="#/completed"
-              className="filter__link"
-              data-cy="FilterLinkCompleted"
-            >
-              Completed
-            </a>
-          </nav> */}
-
-        {/* this button should be disabled if there are no completed todos */}
-        {/* <button
-            type="button"
-            className="todoapp__clear-completed"
-            data-cy="ClearCompletedButton"
-          >
-            Clear completed
-          </button>
-        </footer> */}
       </div>
 
-      {/* DON'T use conditional rendering to hide the notification */}
-      {/* Add the 'hidden' class to hide the message smoothly */}
       <Error
         errorMessage={errorMessage}
         onCloseErrorMessage={() => setErrorMessage('')}
       />
-      {/* <div
-        data-cy="ErrorNotification"
-        className={classNames(
-          'notification is-danger is-light has-text-weight-normal',
-          {
-            hidden: !errorMessage,
-          },
-        )}
-      >
-        <button
-          data-cy="HideErrorButton"
-          type="button"
-          className="delete"
-          onClick={() => setErrorMessage('')}
-        />
-        {errorMessage} */}
-      {/* show only one message at a time */}
-      {/* Unable to load todos
-        <br />
-        Title should not be empty
-        <br />
-        Unable to add a todo
-        <br />
-        Unable to delete a todo
-        <br />
-        Unable to update a todo */}
-      {/* </div> */}
     </div>
   );
 };
